backend/requirements.txt

Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
requests>=2.28.2
playwright>=1.32.1
openai>=0.27.6
gunicorn>=20.1.0
dj-database-url>=1.3.0
